release 1.7.1 of PEcAn

+++ scripts/install_pecan.sh
@@ -147,6 +147,7 @@ case "$OS_VERSION" in
     # sudo apt-get -y install apache2 libapache2-mod-php php libapache2-mod-passenger php-xml php-ssh2 php-pgsql
     sudo apt-get -y install apache2 libapache2-mod-php php php-xml php-ssh2 php-pgsql
     # Ubuntu 14.04 php5-pgsql libapache2-mod-php5 php5 and no php-xml
+    sudo apt-get -y install docker-ce
     ;;
 esac
 
@@ -160,10 +161,12 @@ sudo chmod 755 /usr/local/bin/docker-compose
 sudo usermod -a -G docker carya
 
 cd
-mkdir postgres output portainer
+mkdir -p postgres output portainer
 chmod 777 postgres output portainer
 cd output
-ln -s workflows .
+if [ ! -e workflows ]; then
+  ln -s workflows .
+fi
 cd
 
 cat << EOF > docker-compose.yml
@@ -303,6 +306,7 @@ chmod 755 start-containers.sh
 # load postgres container
 docker-compose pull
 docker-compose -p carya up -d postgres
+sleep 60
 docker run --rm --network carya_pecan pecan/bety:latest initialize
 docker-compose -p carya up -d
 
@@ -392,29 +396,24 @@ sudo cp ed_2.1-opt /usr/local/bin/ed2.git
 echo "######################################################################"
 echo "SIPNET"
 echo "######################################################################"
-if [ ! -e ${HOME}/sipnet_unk ]; then
+if [ ! -e ${HOME}/sipnet ]; then
   cd
-  curl -o sipnet_unk.tar.gz http://isda.ncsa.illinois.edu/~kooper/PEcAn/sipnet/sipnet_unk.tar.gz
-  tar zxf sipnet_unk.tar.gz
-  rm sipnet_unk.tar.gz
+  git clone https://github.com/PecanProject/sipnet.git
+  sed -i 's#$(LD) $(LIBLINKS) \(.*\)#$(LD) \1 $(LIBLINKS)#' ${HOME}/sipnet/Makefile
 fi
-cd ${HOME}/sipnet_unk/
-make clean
-make
-sudo cp sipnet /usr/local/bin/sipnet.runk
-make clean
 
-if [ ! -e ${HOME}/sipnet_r136 ]; then
-  cd
-  curl -o sipnet_r136.tar.gz http://isda.ncsa.illinois.edu/~kooper/PEcAn/sipnet/sipnet_r136.tar.gz
-  tar zxf sipnet_r136.tar.gz
-  rm sipnet_r136.tar.gz
-  sed -i 's#$(LD) $(LIBLINKS) \(.*\)#$(LD) \1 $(LIBLINKS)#' ${HOME}/sipnet_r136/Makefile
-fi
-cd ${HOME}/sipnet_r136/
+cd ${HOME}/sipnet/
+git checkout r136
 make clean
 make
 sudo cp sipnet /usr/local/bin/sipnet.r136
+make clean
+
+cd ${HOME}/sipnet/
+git checkout master
+make clean
+make
+sudo cp sipnet /usr/local/bin/sipnet.git
 make clean
 
 echo "######################################################################"
